[BE] feat: 상담사 이름으로 상담사 목록 조회, 기관 이름으로 기관 목록 조회 기능 구현 완료

See merge request s11-webmobile1-sub2/S11P12B301!80

+++ BE/src/main/java/com/ssafy/inmind/user/dto/ReviewRequestDto.java
@@ -18,6 +18,6 @@ public class ReviewRequestDto {
 
     private String content;
 
-    private float score;
+    private int score;
 
 }
